First Setup + Basic Code
Imported Objects. (Created materials + textures folders to store models'
components; Reorganized the models + materials, as some parts were named
incorrectly; Created some prefabs from the models + Added them to the
scene with a basic layout)
Added waypoints to the board for respected players, both red and blue
(Read further for explanation)
Created a basic script that allows player to move using W and S (Or
anything else that Unity considers the Vertical Axis to be). The player
will move along a set path of waypoints, but can stop anywhere
in-between them and continue in the opposite direction. As the player
reaches a waypoint, the script sets his next and last waypoints to the
following waypoint to reach, and the one he just passed.
The same script contains a Turning section, which is activated by the
Horizontal axis. When a player is turned around (facing the opposite
direction), their forward and backwards code is swapped, so they will
basically move in the direction the are facing. Players have an active
bool, which allows them to move and turn. This will set who the player
is controlling. They also have a blue/red enum, to set which team they
are on. There are 2 speed settings, one for moving (10 by default), and
one for turning (125 by default).
In the next push, I'll try to get the camera to move towards a player as
they become active, and then allow the player to change who they are
controlling.